Create field for signal label, raw label, engineering label for change content label and show inside result groupbox. CalculationService class has been deleted, the math formulas are executed in CaculationApp Class

+++ Model/PlcProfile.cs
@@ -3,6 +3,6 @@
 public class PlcProfile
 {
     public string ProfileName { get; set; } = string.Empty;
-    public int RawMin { get; set; }
-    public int RawMax { get; set; }
+    public double RawMin { get; set; }
+    public double RawMax { get; set; }
 }
